requirements.txt
pypdf
pytesseract
Pillow

langchain
langchain-core
langchain-community

google-generativeai
langchain-google-genai

faiss-cpu

python-dotenv

fastapi
uvicorn
python-multipart

streamlit